requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
folium>=0.14.0
gspread>=5.0.0
requests>=2.31.0

# Genetic Algorithm and Optimization
maturin>=1.0.0  # For building Rust extensions

# Data Processing
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
mypy>=1.0.0
ruff>=0.0.280

# Jupyter support
jupyter>=1.0.0
notebook>=6.0.0
ipykernel>=6.0.0

# Performance monitoring
memory-profiler>=0.60.0
psutil>=5.9.0

# Optional: For additional geocoding features
geopy>=2.3.0

# Environment variable management
python-dotenv>=1.0.0 